+++ datatypes-cose/src/commonMain/kotlin/at/asitplus/crypto/datatypes/cose/CoseAlgorithm.kt
@@ -49,8 +49,26 @@ enum class CoseAlgorithm(val value: Int) {
         RS384 -> JwsAlgorithm.RS384
         RS512 -> JwsAlgorithm.RS512
     }
-}
 
+}
+fun JwsAlgorithm.toCoseAlgorithm() = when (this) {
+    JwsAlgorithm.ES256 -> CoseAlgorithm.ES256
+    JwsAlgorithm.ES384 -> CoseAlgorithm.ES384
+    JwsAlgorithm.ES512 -> CoseAlgorithm.ES512
+
+    JwsAlgorithm.HS256 -> CoseAlgorithm.HS256
+    JwsAlgorithm.HS384 -> CoseAlgorithm.HS384
+    JwsAlgorithm.HS512 -> CoseAlgorithm.HS512
+
+    JwsAlgorithm.PS256 -> CoseAlgorithm.PS256
+    JwsAlgorithm.PS384 -> CoseAlgorithm.PS384
+    JwsAlgorithm.PS512 -> CoseAlgorithm.PS512
+
+    JwsAlgorithm.RS256 -> CoseAlgorithm.RS256
+    JwsAlgorithm.RS384 -> CoseAlgorithm.RS384
+    JwsAlgorithm.RS512 -> CoseAlgorithm.RS512
+    JwsAlgorithm.NON_JWS_SHA1_WITH_RSA -> TODO()
+}
 
 object CoseAlgorithmSerializer : KSerializer<CoseAlgorithm> {